[chore][pkg/stanza] Move some reader tests into reader package
